First major update of the server
Server expects all websites to be stored in a 'configs' folder in the
root directory of the app. All files within it are expected to be JSON
files which contain: A name (UNIQUE), domain, tld, and an array of subdomains (UNIQUE).

There's still a lot to be added like a valitator, support for multiple
subdomains, etc. But until then this will suffice (...once tested that
is.)